requirements.txt
asgiref==3.5.2
class-registry==2.1.2
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
filters==1.3.2
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.8.17
six==1.16.0
sqlparse==0.4.2
